Update Aotearoa Treasures’ inventory management system
I've updated the main.cpp and made it so you can save the stock and staff updates and removals before it was just preloaded and would reset the updates but now it should be working. It still has the preloaded staff only if the staff file and stock files are empty.